# [Differ] Remove DefaultDiffer, use ConsoleDiffer instead to avoid double diffing process