--- a/phpstan.neon
+++ b/phpstan.neon
@@ -26,6 +26,8 @@ parameters:
         - tests
         - rules-tests
         - utils
+        - e2e/e2eTestRunnerWithCache.php
+        - e2e/e2e/e2eTestRunner.php
 
     scanDirectories:
         - stubs