… src/Console/Formatter/ConsoleDiffer.php