@staabm this is the rework effort from the old PR:

- https://github.com/rectorphp/rector-src/pull/3705
- https://github.com/rectorphp/rector-src/pull/6482

I tested with diff on json format, it keep same, see https://www.diffchecker.com/YtVb6oKF/

The process is:

- Use `DefaultDiffer->diff()` as default value
- When pass to `FileDiff::__construct()` on `diffConsoleFormatted`, it call

```php
$this->colorConsoleDiffFormatter->format($diff)
```

like this:

```php
        $diff = $this->defaultDiffer->diff($oldContent, $newContent);

        // always keep the most recent diff
        return new FileDiff(
            $relativeFilePath,
            $shouldShowDiffs ? $diff : '',
            $shouldShowDiffs ? $this->colorConsoleDiffFormatter->format($diff) : '',
            $rectorsWithLineChanges
        );
```

Then, to keep the string `@@ @@`, I use regex on `ColorConsoleDiffFormatter->diff()`.