@samsonasik 2025-09-21
Not yet check on a large diff. The check may need a large diff on a single run to see comparison, yes?