@samsonasik 2025-09-21
Updated stop loop on found --- Original and +++ New, got new result:


**Before**

```
bin/rector --clear-cache --dry-run  8.33s user 0.09s system 99% cpu 8.441 total
```

**After**

```
bin/rector --clear-cache --dry-run  6.93s user 0.09s system 99% cpu 7.033 total
```