@TomasVotruba 2025-09-21
@samsonasik What are numbers on one if our private client's project? Real projects are better than Rector itself