@samsonasik 2025-09-21
@TomasVotruba yes, the measure is depends on how many files changed, more diffs, more faster.

I tried in our project with **70 files changed**, when there are diffs, it 3 seconds faster.

**Before**

```
../rector-src/bin/rector process src --dry-run --clear-cache  796.63s user 21.67s system 652% cpu 2:05.41 total
```

**After**

```
../rector-src/bin/rector process src --dry-run --clear-cache  772.05s user 20.86s system 647% cpu 2:02.38 total
```